refactor: move GPU compute utilities into gpu/ directory
gpu.rs becomes gpu/mod.rs (GpuContext path unchanged), with 4 compute
utilities as sub-modules:

- gpu/nv12_converter.rs - RGBA to NV12 compute shader + triple-buffer
- gpu/rgba_readback.rs - RGBA triple-buffered staging readback
- gpu/color_grade.rs - 1D LUT color grading pass
- gpu/yuv_stack_packer.rs - YUV tile packing compute shader

Shader include_str! paths updated for new directory depth.
All import paths updated across the workspace. No re-exports.
