+++ crates/reco-core/src/gpu/mod.rs
@@ -487,3 +487,8 @@ mod tests {
         let _ = reconstructed.queue();
     }
 }
+
+pub mod color_grade;
+pub mod nv12_converter;
+pub mod rgba_readback;
+pub mod yuv_stack_packer;
