+++ crates/reco-core/src/session/mod.rs
@@ -52,8 +52,8 @@ use crate::async_encode::AsyncEncodeThread;
 use crate::core::StitchCore;
 use crate::core::types::StitchCoreConfig;
 use crate::detect::director::ViewportPosition;
+use crate::gpu::nv12_converter::Nv12Converter;
 use crate::gpu::{GpuContext, OutputFormat};
-use crate::nv12_converter::Nv12Converter;
 use crate::render::pipeline::StitchPipeline;
 use crate::render::renderer::InputFormat;
