+++ crates/reco-core/src/session/types.rs
@@ -8,8 +8,8 @@ use crate::calibration::MatchCalibration;
 use crate::core::types::StitchCoreError;
 use crate::detect::director::{MappedDetection, ViewportPosition};
 use crate::encoder::{EncodeError, Encoder};
+use crate::gpu::nv12_converter::Nv12Error;
 use crate::gpu::{GpuContext, GpuError, OutputFormat};
-use crate::nv12_converter::Nv12Error;
 use crate::render::pipeline::PipelineError;
 use crate::render::renderer::InputFormat;
 use crate::render::viewport::ViewportConfig;
